Added Lane change rate calc on  per car basis
Also added capability to plot the lane id over time for a selected car.
With this commit, there are two ways of calculating lane change rate.

Method 1: Implemented in the previous commit to this
Compare lane id of every car at timestep t+1 to that same cars lane id
at timestep t. If there is a change, increment lane change by 1. Finally
divide that lane change number by 100 aka number of agents

Method 1 captures every lane change made by every car. So there is every
possiblity that the lane change rate will be a number higher than 1. See
corresponding discussion for method 2 at the end of this commit message.

Method 2: Implemented in this commit
Look at the lane id of a car over the entire 200 timesteps. Even if the
lane id has changed once, then add 1 to the number of lane changes. Do
this for every car. Divide by 100 to get the lanechangerate for that 1
simulation. Remember that there are 100 such simulations.

Method 2 will give a much lower number than method 1 because all it is
calculating is how many cars in the 100 changed their lane over the
entire 20 second simulation. This number will always be lesser than
equal to 100 (since there are 100 cars in this situation). Therefore the
lane change rate will always be lesser than or equal to 1.