Working without laneid feature for RAIL policies
Motivation: Calculate lane change rate as an emergent metric for
policies used to demostrate reward augmented imitation learning. These
policies were before laneid was included as a feature so they work with
the set of 66 features instead of 67.

Ideally this should have been done from the master branch but have been
developing on laneid branch so continued with that. In a recent commit,
added lane number as being tracked in the step_infos within the
multiagent_ngsim_env.jl file.