Extraction of ngsim features car wise
To be able to compute lane change rate and timegap looking at
trajectories of cars one by one instead of everything bundled together
which is how the collision rate, offroad and hardbrake rates were being
measured so far. The bundling approach worked for those metrics because
they are binary. Not applicable to timegap and lane change rate
calculation.