Fixed seed pure accl from Multilane to drive ego
Converting Multilane.MLAction into AccelTurnrate to work with the
propagate function from AutomotiveDrivingModels called within the _step!
function.
Note that there is no planning done yet. Just using a fixed random seed
that generates the same action everytime and that too the action is a
pure acceleration with the second component being 0.

Files modified:
multiagent_ngsim_env_videomaker.jl
- Converting MLAction to AccelTurnrate by accessing the acc and ydot
  fields within the MLAction struct
- Hard coding the render method with the vehicle id being controlled by
  MLAction to color it yellow

Issues: The video shows the yellow car to be turning even though it
should only be accelerating as the random seed was chosen to be 15 which
yields accl of 1.0 and nothing on the ydot